Harrison 2646 - KJON/PMCL - SWAPI: use sci_start_time from L2 instead of epoch
sci_start_time is the start of the sweep, while epoch is the midpoint.
We had assumed that epoch was the start of the sweep.
Also, when calculating proton peak angle, add a half-bin correction to get the time at the center of the peak bin instead of the start.